requirements.txt
langchain
langchain-community
openai
pinecone-client
python-dotenv
pypdf
tqdm
requests
fastapi
uvicorn[standard]